NMS-8301: Move BusinessServiceRestService to its own package
