+++ opennms-webapp-rest/src/main/webapp/WEB-INF/applicationContext-cxf-rest-v2.xml
@@ -29,7 +29,7 @@
     -->
     <bean id="springResourceContextProvider" class="org.opennms.web.rest.support.SpringResourceContextProvider"/>
 
-    <jaxrs:server id="cxf-rest-v2" address="/" basePackages="org.opennms.web.rest.v2">
+    <jaxrs:server id="cxf-rest-v2" address="/" basePackages="org.opennms.web.rest.v2,org.opennms.web.rest.v2.bsm">
       <jaxrs:properties>
         <!-- Use the ResourceContextProvider defined above -->
         <entry key="org.apache.cxf.jaxrs.resource.context.provider" value-ref="springResourceContextProvider"/>
